# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Custom Printing E-commerce API
  version: 1.0.0
  description: >-
    OpenAPI specification generated from JSDoc annotations in the Express
    codebase.
servers:
  - url: http://localhost:3002
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/v1/admin/coupons:
    get:
      summary: Get all coupons
      description: Admin can view all coupons with usage statistics
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of coupons retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        code:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                          nullable: true
                        discountType:
                          type: string
                          enum:
                            - PERCENTAGE
                            - FIXED
                        discountValue:
                          type: number
                        minPurchaseAmount:
                          type: number
                          nullable: true
                        maxDiscountAmount:
                          type: number
                          nullable: true
                        usageLimit:
                          type: integer
                          nullable: true
                        usageLimitPerUser:
                          type: integer
                        validFrom:
                          type: string
                          format: date-time
                        validUntil:
                          type: string
                          format: date-time
                        isActive:
                          type: boolean
                        applicableTo:
                          type: string
                          enum:
                            - ALL
                            - CATEGORY
                            - PRODUCT
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                        _count:
                          type: object
                          properties:
                            usages:
                              type: integer
        '401':
          description: Unauthorized - Admin authentication required
    post:
      summary: Create new coupon
      description: Admin can create a new discount coupon
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name
                - discountType
                - discountValue
                - validFrom
                - validUntil
              properties:
                code:
                  type: string
                  example: SAVE20
                name:
                  type: string
                  example: Save 20% Off
                description:
                  type: string
                  example: Get 20% off on all products
                discountType:
                  type: string
                  enum:
                    - PERCENTAGE
                    - FIXED
                  example: PERCENTAGE
                discountValue:
                  type: number
                  example: 20
                minPurchaseAmount:
                  type: number
                  nullable: true
                  example: 1000
                maxDiscountAmount:
                  type: number
                  nullable: true
                  example: 500
                usageLimit:
                  type: integer
                  nullable: true
                  example: 100
                usageLimitPerUser:
                  type: integer
                  example: 1
                validFrom:
                  type: string
                  format: date-time
                  example: '2024-01-01T00:00:00Z'
                validUntil:
                  type: string
                  format: date-time
                  example: '2024-12-31T23:59:59Z'
                isActive:
                  type: boolean
                  example: true
                applicableTo:
                  type: string
                  enum:
                    - ALL
                    - CATEGORY
                    - PRODUCT
                  example: ALL
      responses:
        '200':
          description: Coupon created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - message
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Coupon created successfully
                  data:
                    type: object
        '400':
          description: Validation error or coupon code already exists
        '401':
          description: Unauthorized - Admin authentication required
...